diff - new opt LYD_DIFF_DUPINST_FULL none data
In the case of key-less lists or non-configuration leaf-lists, a diff is
reported only with the changed items.

Sometimes, it is useful to know the latest state of the list along with
individual changes.

One specific usecase is for gNMI Path construction, as positional
predicates in paths cannot be converted into a gNMI Path.
So, leaflist changes are treated as a replace of the entire leaflist.

So, when this option is set, all leaflist nodes in the second data tree
are returned with operation none.